Upload documents to S3
After discussing with CS, I decided to change to the more canonical
aws-sdk from the s3 gem. The workflow for the sdk is somewhat more
complicated, but it seemed sensible to use the actual AWS gem.

The app has also been converted from classic sinatra to modular
sinatra. This makes it easier to group the functionality and, more
importantly, makes mutation testing much simpler.