Change README references and add support files
DS has created a separate repo of the scripts used to provision S3
buckets. As per CS’s comment on the PR, I have changed the README to
reflect this (and will not be including the scripts from the prototype).
Also, checking over the README, I realised I had forgotten to add the
template `.env.example` file earlier.