Stop the uploader breaking in production (#11)
* Stop the uploader breaking in production

As the keys are no longer required due to our use of roles.  This commit
brings it in line with the downloader.

* Make the logs kibana-friendly

* Stub IAM role response with example from AWS docs

I was going to record this with VCR, but the response was simpler than I
expected it to be, so I stubbed it with webmock instead. See the
comments for a link to the AWS documentation that the example was taken
from.

* Fix the broken expectation

I believe the change to Zulu time is caused by the fact that the stubbed
IAM credentials request returns Zulu time.  As this example was taken
straight from the AWS docs, I have changed the expectation to reflect
the change rather than the changing the AWS example to use UTC.

* Stop the logs appearing in the test runs

They were also coming back in the wrong format (Apache, not logstash). I
will look into this later.

* typo: appication -> application