Expose status of S3 from AWS RSS feed
As a subkey of dependencies, with a keyname taken from the bukcet
region. This should make it explicit which bucket region is being used.
This is the AWS reported status of the region, only.  Something could
still be wrong with the bucket (permissions, etc.), which will be
covered by a different spec.