Healthcheck showing s3 (#17)
* Expose status of S3 from AWS RSS feed

As a subkey of dependencies, with a keyname taken from the bukcet
region. This should make it explicit which bucket region is being used.
This is the AWS reported status of the region, only.  Something could
still be wrong with the bucket (permissions, etc.), which will be
covered by a different spec.

* Avoid leaking info/make test debugging easier

The standard Sintara error pages are verbose and give away alot of
information.  Disabling the error pages and raising errors to STDERR
closes one possible vector of attack and makes it much easier to find
the actual error in failing request specs.

* Reduce maintainence overheads with contstants

The status rss feed should not change in the foreseeable future.  This
makes it simpler to maintain.

* Clean up dotenv example

The .env.example version was out-of-date and should have been removed a
while ago.  Also added the AWS_REGION variable to the up-to-date
version.

* Make it simpler to manage test ENV variables

Scoped to `spec_helper.rb` only to prevent it from becoming a problem
with deployment.

* Kill some mutations

That were not easy to target using the feature spec that has been the
main driver for this commit, otherwise

* Make the feature spec scope clearer

A single healthcheck feature spec is in danger of growing VERY long.