Ensure AV dependency is working as expected
A healthcheck will trigger two calls to the AV: one with an eicar test
virus and one without.  I considered sending these using a timed job and
saving the results between runs.  I rejected that approach because of
the added complexity and the expected light loading of the service. It
can always be added later if required.