Healthcheck av (#18)
* Ensure the specs get found

Ahem.  Stupid naming oversight in the last PR.

* Make service status info simpler to maintain

No need for both a `/healthcheck` and a `/status` endpoint.  I've
consolidated them both into `/healthcheck`. Honestly, this probably does
not need a `service_status` key, either.  However, I moved it in from
`/status` for the sake of consistency.

* Ensure AV dependency is working as expected

A healthcheck will trigger two calls to the AV: one with an eicar test
virus and one without.  I considered sending these using a timed job and
saving the results between runs.  I rejected that approach because of
the added complexity and the expected light loading of the service. It
can always be added later if required.