Healthcheck av (#19)
* Ensure the specs explain what is going on

When run using `--format documentation`.

Coming back from the Christmas holiday, I couldn't remember where I was
with all of this.  Reading and running the specs didn't help much, so I
thought it was time I clarified the ones I was working on to ensure they
made the workflow clear.  Eventually, I would like to do all of them,
but that is outside the scope of this excercise.

* Clairify method names, fail on dependency failures

The healthcheck method names were confusing, particularly for the
av.checks.  I've simplified these and simplifed the overall workflow,
somewhat. I have also configured the checks such that a failure in a
dependency will cause the app healthcheck to fail.

* Make the purpose of this key clear