Allow for the addition of a subfolder
This is distinct from JL's branch `subfolders-spike`   The intetion here
is to get the minimum amount of code in place needed to make the full
stack work with subfolders.  That said, the test suite here still passes
with minimal changes and the downloader, at least, should continue to
work without significant alterations (see `subfolders-spike` branch on
that project).