Ensure we can diagnose upload failures
We were not logging previously on this app.  That made it difficult to
diagnose the intermittent connection failures we see with S3 in some
environments.  This implements logging for uploads and for the
healthcheck write test (which uses the same upload method in the
background), which should give us some insight into the connection
problems.