Ensure we can diagnose upload failures (#31)
* Ensure we can diagnose upload failures

We were not logging previously on this app.  That made it difficult to
diagnose the intermittent connection failures we see with S3 in some
environments.  This implements logging for uploads and for the
healthcheck write test (which uses the same upload method in the
background), which should give us some insight into the connection
problems.

* Remove unnecessary timestamp key/value pair

LogStashLogger includes a timestamp by default.

* Allow parsing by LogStashLogger severity levels

By default LogStashLogger includes a `@severity` key that can (should)
be used to reveal whether or not the request succeded or failed.  This
replaces the `success: true|false` key.

* Make it easy to debug errors from the logs

The default formatter, `:json_line`, is compact and difficult to read.
This expands the formatting by pretty-printing the json, which should
not change its parsability, but will make it much easier for us to read.