Make the uploader more tolerant of s3 errors
While the exact nature of the errors are not currently clear, it seems
likely that they result from a combination of S3 checksumming errors and
the small number of retries allocated to resolve these.  My reasoning is
that it has been reported that, despite the errors, the files are often
successfully uploaded.

This commit turns off all unnecessary checksumming and increases the
number of retires from three to five.  Hopefully, these combined
changes will resolve the issue.