Resolve sentry issues (#45)
* Ensure the errors make it to sentry

The `false` being returned was preventing this from happening.
Re-raising them should do it.

* Ensure the new error workflow on add is handled

The specs were breaking and it also seemed prudent to return the error
as a json reponse to the client that made the request. Prior to this
checkin, the TaxTribsDatacapture failure was not completely clear as its
JSON parsing was breaking on the html error page this app returned.

This way, we should be able to do some diagnosis of the error in Sentry
even if the uploader fails to send it; the message will be part the json
response message recorded for TaxTribsData.

* Remove one of the common timeout failure cases

This simply checks the AWS status page and reports errors.  We aren't
using it and it times out regularly.